refactor: cookie config
