+++ packages/app-builder/src/utils/preferences-cookies/preferences-cookies-write.ts
@@ -1,8 +1,6 @@
 import Cookie from 'js-cookie';
 
-import { type PreferencesCookie, PreferencesCookieSchema } from './types';
-
-const COOKIE_NAME = 'preferences';
+import { COOKIE_NAME, type PreferencesCookie, PreferencesCookieSchema } from './config';
 
 export function setPreferencesCookie<K extends keyof PreferencesCookie>(
   key: K,
@@ -27,10 +25,24 @@ export function setPreferencesCookie<K extends keyof PreferencesCookie>(
     throw new Error('Invalid preferences cookie value');
   }
 
-  current[key] = value ? '1' : '0';
+  // Store the value as a string, serializing if necessary
+  switch (typeof value) {
+    case 'boolean':
+      current[key] = value ? '1' : '0';
+      break;
+    case 'string':
+      current[key] = value;
+      break;
+    case 'number':
+      current[key] = (value as number).toString();
+      break;
+    default:
+      current[key] = JSON.stringify(value);
+      break;
+  }
 
   Cookie.set(COOKIE_NAME, JSON.stringify(current), {
     expires: 365,
-    sameSite: 'lax',
+    sameSite: 'strict',
   });
 }
